Refactor: "레벨 디자인 정수형 -> 실수형으로 변경"

+++ Assets/00_Scripts/Scriptable/LevelDesign.cs
@@ -17,15 +17,15 @@ public class LevelDesign : ScriptableObject
 public class LevelData
 {
     [Range(0.0f, 10.0f)]
-    public int C_ATK, C_HP, C_EXP, C_MAXEXP, C_MONEY; // 현재 레벨의 공격력, 체력, 경험치, 최대 경험치, 돈
+    public float C_ATK, C_HP, C_EXP, C_MAXEXP, C_MONEY; // 현재 레벨의 공격력, 체력, 경험치, 최대 경험치, 돈
 
     [Space(20)]
     [Header("## Base Value")]
-    public int B_ATK;
-    public int B_HP;
-    public int B_EXP;
-    public int B_MAXEXP;
-    public int B_MONEY;
+    public float B_ATK;
+    public float B_HP;
+    public float B_EXP;
+    public float B_MAXEXP;
+    public float B_MONEY;
 
     public double ATK() => Utils.CalculatedValue(B_ATK, Base_Manager.Data.Level, C_ATK);
     public double HP() => Utils.CalculatedValue(B_HP, Base_Manager.Data.Level, C_HP);
@@ -38,12 +38,12 @@ public class LevelData
 public class StageData
 {
     [Range(0.0f, 10.0f)]
-    public int M_ATK, M_HP, M_MONEY;
+    public float M_ATK, M_HP, M_MONEY;
     [Space(20f)]
     [Header("## Base Value")]
-    public int B_ATK;
-    public int B_HP;
-    public int B_MONEY;
+    public float B_ATK;
+    public float B_HP;
+    public float B_MONEY;
 
     public double ATK() => Utils.CalculatedValue(B_ATK, Base_Manager.Data.Stage, M_ATK);
     public double HP() => Utils.CalculatedValue(B_HP, Base_Manager.Data.Stage, M_HP);
